requirements.txt
click>=8.1
pandas>=2.1
requests>=2.31
google-play-scraper>=1.2
google-genai>=0.4
python-dotenv>=1.0
pydantic>=2.0
PyYAML>=6.0
scikit-learn>=1.3
